Plant Tracker Upload Form
Added a proper django form to deal with gpx file uploads to be added to
the map. Altered the logic for dealing with uploaded files and
refactored file processing to take place in its own function.

Made minor improvements to the styling of the buttons associated with
the file upload form.

Signed-off-by: Avery Uslaner <uslaner.avery@gmail.com>