Plant Database Connection Removed
Removed separate plant database connection upon realization that the minor benefit of having plant database info in its own database would not be worth the hassle.

Plant database records need to be linked to the User model in the default database which isn't supported by Django and would require some hacky workarounds that would ultimately harm the maintainability of the app.

Signed-off-by: Avery Uslaner <ctsadmin@conservationtechnologysolutions.org>