Plant Database Accession Form, View, Controller
Added new page with a form for users to create new accession objects.

Added Accession objects to the admin view.

Updated urls to include the new accession form.

Added javascript to facilitate the addition of multiple collectors when filling out the accession creation form.

Updated the plant database index page to display a list of accessions in the database. When empty, the index page only displays a link to the add accession form.

Signed-off-by: Avery Uslaner <ctsadmin@conservationtechnologysolutions.org>