Travis Updates
Removed python 3.5 from Travis since it is not compatible with Django 3.0.

Added python 3.7 to check if that version will be compatible with the current environment.

Specified Bionic beaver build environment in an attempt to address an error with an out of date GDAL version which was throwing an error during testing.

Signed-off-by: Avery Uslaner <ctsadmin@conservationtechnologysolutions.org>