Changes to Seed Accession Addition Form
Added a field for choose Altitude unit from a defined list of either feet or meters.

Converted family and genus form fields from ModelChoiceField to Charfield since it was too problematic to offer both a list of options and the ability for users to add new options especially since it would be difficult to provide an up-to-date fixture of plant families and genera.

Conversely, the country form field was converted from a Charfield to a ModelChoiceField since it's more practical to provide and manage a fixture of all country names.

The bank_date field was also changed to use a SelectDateWidget which is a bit more user friendly.

Signed-off-by: Avery Uslaner <ctsadmin@conservationtechnologysolutions.org>